requirements.txt
# Core ML Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data Processing
openpyxl>=3.1.0  # For Excel file support

# Visualization (optional, for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
PyYAML>=6.0

# Jupyter (for notebooks)
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.25.0

joblib
mlflow
